+++ main.lua
@@ -17,7 +17,24 @@ require("skillPickup")
 require("drop_item")
 require("compat_patch")
 
+mods["MGReturns-ENVY"].auto()
+envy = mods["MGReturns-ENVY"]
+public_things = {
+    ["drop_item"] = function(player, item_id, item_object_id)
+        return drop_item(player, item_id, item_object_id)
+    end,
+    ["Utils"] = Utils,
+    ["skill_create"] = function(x, y, skill_params)
+        return skill_create(x, y, skill_params)
+    end
+} -- Maybe using a wrong way
+require("./envy_setup")
+
 local tooltip
+
+gm.post_script_hook(gm.constants.ui_hover_tooltip, function(self, other, result, args)
+    tooltip = args[3].value
+end)
 
 gui.add_always_draw_imgui(function()
     if ImGui.IsKeyPressed(params['drop_item_key'], false) then
@@ -31,12 +48,11 @@ gui.add_always_draw_imgui(function()
                             drop_item(player, item_id, item_object_id)
                         end
                     else
-                        local skill =
-                        Utils.find_skill_with_localized(tooltip, player)
+                        local skill = Utils.find_skill_with_localized(tooltip, player)
                         if skill.skill_id ~= nil and skill.slot_index ~= nil and skill.skill_id ~= 0 then
                             if skill_create ~= nil then
                                 gm.actor_skill_set(player, skill.slot_index, 0)
-                                skill_create(player, skill)
+                                skill_create(player.x, player.y, skill)
                             end
                         end
                     end