+++ skillPickup.lua
@@ -1,14 +1,11 @@
-require("Utils")
 local function setupSkill(target, skill_params)
-    gm._mod_instance_set_sprite(target, skill_params.skill_sprite)
-    target.image_index = skill_params.skill_subimage
-    target.translation_key = skill_params.skill_translation_key
-    target.skill_id = skill_params.skill_id
-    target.slot_index = skill_params.slot_index
+    for k,v in pairs(skill_params) do
+        target[k] = v
+    end
     target.text = gm.ds_map_find_value(Utils.get_lang_map(), target.translation_key .. ".name")
 end
-local drop_skill_send, activate_skill_send, drop_skill
-skill_create = function(player, skill_params)
+local drop_skill_send, activate_skill_send, drop_skill, activate_skill
+skill_create = function(x, y, skill_params)
     log.error("skill_create hasn't been initialized")
 end
 local function init()
@@ -16,7 +13,7 @@ local function init()
     active_skill_packet:onReceived(function(message, player)
         local Player = message:read_instance().value
         local Interactable = message:read_instance().value
-        if Net.get_type() == Net.TYPE.host then
+        if Utils.get_net_type() == Net.TYPE.host then
             local sync_message = active_skill_packet:message_begin()
             sync_message:write_instance(Player)
             sync_message:write_instance(Interactable)
@@ -29,13 +26,7 @@ local function init()
         local sync_message = active_skill_packet:message_begin()
         sync_message:write_instance(Player)
         sync_message:write_instance(Interactable)
-        if Net.get_type() == Net.TYPE.host then
-            sync_message:send_to_all()
-        else
-            sync_message:send_to_host()
-        end
-        gm.actor_skill_set(Player, Interactable.slot_index, Interactable.skill_id)
-        gm.instance_destroy(Interactable.id)
+        return sync_message
     end
 
     local skillPickup = Interactable.new("hinyb", "skillPickup")
@@ -46,15 +37,17 @@ local function init()
                               .active_skill
             if skill.skill_id ~= 0 then
                 gm.actor_skill_set(Player.value, skill.slot_index, 0)
-                skill_create(Player.value, {
+                skill_create(Player.value.x, Player.value.y, {
                     slot_index = skill.slot_index,
                     skill_id = skill.skill_id,
-                    skill_sprite = skill.sprite,
-                    skill_translation_key = string.sub(skill.name, 1, -6),
-                    skill_subimage = skill.subimage
+                    sprite_index = skill.sprite,
+                    translation_key = string.sub(skill.name, 1, -6),
+                    image_index = skill.subimage
                 })
             end
-            activate_skill_send(Player.value, Interactable.value)
+            activate_skill (Player, Interactable)
+            gm.actor_skill_set(Player.value, Interactable.value.slot_index, Interactable.value.skill_id)
+            Interactable:destroy()
         end
     end)
 
@@ -62,17 +55,17 @@ local function init()
     drop_skill_packet:onReceived(function(message, player)
         local slot_index = message:read_int()
         local skill_id = message:read_int()
-        local skill_sprite = message:read_int()
-        local skill_translation_key = message:read_string()
-        local skill_subimage = message:read_int()
+        local sprite_index = message:read_int()
+        local translation_key = message:read_string()
+        local image_index = message:read_int()
         local skill_params = {
             slot_index = slot_index,
             skill_id = skill_id,
-            skill_sprite = skill_sprite,
-            skill_translation_key = skill_translation_key,
-            skill_subimage = skill_subimage
+            sprite_index = sprite_index,
+            translation_key = translation_key,
+            image_index = image_index
         }
-        if Net.get_type() == Net.TYPE.host then
+        if Utils.get_net_type() == Net.TYPE.host then
             drop_skill(player.x, player.y, skill_params)
         else
             local skill = message:read_instance().value
@@ -83,35 +76,44 @@ local function init()
         local sync_message = drop_skill_packet:message_begin()
         sync_message:write_int(skill_params.slot_index)
         sync_message:write_int(skill_params.skill_id)
-        sync_message:write_int(skill_params.skill_sprite)
-        sync_message:write_string(skill_params.skill_translation_key)
-        sync_message:write_int(skill_params.skill_subimage)
-        if skill ~= nil then
-            sync_message:write_instance(skill)
+        sync_message:write_int(skill_params.sprite_index)
+        sync_message:write_string(skill_params.translation_key)
+        sync_message:write_int(skill_params.image_index)
+        return sync_message
+    end
+    drop_skill = function(x, y, skill_params)
+        local skill = gm.instance_create(x - 20, y - 20, skillPickup.value)
+        setupSkill(skill, skill_params)
+        gm.call("gml_Script_interactable_sync", skill, skill)
+        local sync_message = drop_skill_send(skill_params, skill)
+        sync_message:write_instance(skill)
+        sync_message:send_to_all()
+    end
+    gm.post_script_hook(gm.constants.run_create, function(self, other, result, args)
+    if Utils.get_net_type() == Net.TYPE.single then
+        skill_create = function(x, y, skill_params)
+            local skill = gm.instance_create(x - 20, y - 20, skillPickup.value)
+            setupSkill(skill, skill_params)
+        end
+        activate_skill = function (Player, Interactable) end
+    elseif Utils.get_net_type() == Net.TYPE.host then
+        skill_create = function(x, y, skill_params)
+            drop_skill(x, y, skill_params)
+        end
+        activate_skill = function (Player, Interactable)
+            local sync_message = activate_skill_send(Player, Interactable)
             sync_message:send_to_all()
-        else
+        end
+    else
+        skill_create = function(x, y, skill_params)
+            local sync_message = drop_skill_send(skill_params)
+            sync_message:send_to_host()
+        end
+        activate_skill = function (Player, Interactable)
+            local sync_message = activate_skill_send(Player, Interactable)
             sync_message:send_to_host()
         end
     end
-    drop_skill = function(x, y, skill_params)
-        local skill = gm.instance_create(x, y, skillPickup.value)
-        gm.call("gml_Script_interactable_sync", skill, skill)
-        setupSkill(skill, skill_params)
-        drop_skill_send(skill_params, skill)
-    end
-end
-local function update_multiplayer(host)
-    if host then
-        skill_create = function(player, skill_params)
-            drop_skill(player.x, player.y, skill_params)
-        end
-    else
-        skill_create = function(player, skill_params)
-            drop_skill_send(skill_params)
-        end
-    end
-end
-gm.post_script_hook(gm.constants.update_multiplayer_globals, function(self, other, result, args)
-    update_multiplayer(args[2].value)
 end)
+end
 Initialize(init)